Moved cucumber code from main to test directory
