+++ bddtest/src/test/java/bddtest/elasticsearch/client/DeleteIndex.java
@@ -1,13 +1,8 @@
 package no.dcat.bddtest.elasticsearch.client;
 
-import no.dcat.harvester.crawler.handlers.ElasticSearchResultHandler;
 import no.difi.dcat.datastore.Elasticsearch;
-import org.elasticsearch.action.ActionFuture;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
-import org.elasticsearch.action.admin.indices.delete.DeleteIndexResponse;
-import org.elasticsearch.action.bulk.BulkRequestBuilder;
 import org.elasticsearch.index.IndexNotFoundException;
-import org.elasticsearch.index.query.QueryBuilders;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
